feat: 수정 및 생성 페이지에서 원문 보기 버튼 및 반응형 작업
