+++ src/shared/components/OriginView.tsx
@@ -28,7 +28,7 @@ export const OriginView = ({
         width="100%"
         height="100%"
         title="원본 보기"
-        className={`${className} hide-scrollbar`}
+        className={`${className}`}
       />
     </div>
   );
